fix gradle check under jigsaw
closes #14726

Squashed commit of the following:

commit 5b591e98570e3fa481b2816a44063b98bff36ddf
Author: Robert Muir <rmuir@apache.org>
Date:   Fri Nov 13 00:54:08 2015 -0500

    add assumption for self-signing in PluginManagerTests

commit ed11e5371b6f71591dc41c6f60d033502cfcf029
Author: Robert Muir <rmuir@apache.org>
Date:   Fri Nov 13 00:20:59 2015 -0500

    show error output from integ test startup

commit d8b187a10e95d89a0e775333dcbe1aaa903fb376
Author: Robert Muir <rmuir@apache.org>
Date:   Thu Nov 12 22:14:11 2015 -0500

    fix gradle check under jigsaw

+++ qa/evil-tests/src/test/java/org/elasticsearch/plugins/PluginManagerTests.java
@@ -661,9 +661,14 @@ public class PluginManagerTests extends ESIntegTestCase {
 
         SSLSocketFactory defaultSocketFactory = HttpsURLConnection.getDefaultSSLSocketFactory();
         ServerBootstrap serverBootstrap = new ServerBootstrap(new NioServerSocketChannelFactory());
-        SelfSignedCertificate ssc = new SelfSignedCertificate("localhost");
+        SelfSignedCertificate ssc = null;
 
         try {
+            try {
+                ssc = new SelfSignedCertificate("localhost");
+            } catch (Exception e) {
+                assumeNoException("self signing shenanigans not supported by this JDK", e);
+            }
 
             //  Create a trust manager that does not validate certificate chains:
             SSLContext sc = SSLContext.getInstance("SSL");
@@ -700,7 +705,9 @@ public class PluginManagerTests extends ESIntegTestCase {
         } finally {
             HttpsURLConnection.setDefaultSSLSocketFactory(defaultSocketFactory);
             serverBootstrap.releaseExternalResources();
-            ssc.delete();
+            if (ssc != null) {
+                ssc.delete();
+            }
         }
     }
